# docs/swagger.yaml
definitions:
  api.AccountInfo:
    properties:
      balance:
        description: The balance in layer 1
        type: number
      dataTotal:
        description: Total Size of storage data
        type: integer
      fileCount:
        description: Total number of files
        type: integer
      rewardCount:
        description: Total number of distributed reward recodes
        type: integer
      storageFeeTotal:
        description: Total storage fee
        type: number
      txCount:
        description: Total number of layer1 transaction
        type: integer
    type: object
...
  api.AddressStatList:
    description: Hex40 address stat list
    properties:
      list:
        description: Stat list
        items:
          $ref: '#/definitions/api.AddressStat'
        type: array
      total:
        description: The total number of stat returned
        type: integer
    type: object
  api.BusinessError:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
...
  api.RewardList:
    description: Miner reward list
    properties:
      list:
        description: Miner reward list
        items:
          $ref: '#/definitions/api.Reward'
        type: array
      total:
        description: The total number of miner reward returned
        type: integer
    type: object
...
  api.StorageTxList:
    description: Submission information list
    properties:
      list:
        description: Submission list
        items:
          $ref: '#/definitions/api.StorageTxInfo'
        type: array
      total:
        description: The total number of submission returned
        type: integer
    type: object
...
info:
  contact: {}
  description: Use any http client to fetch data from the 0G Storage Scan.
  title: 0G Storage Scan API
  version: "1.0"
paths:
  /accounts/{address}:
    get:
      consumes:
      - application/json
      description: Query account information for specified account
      parameters:
      - description: The account address
        in: path
        name: address
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/api.BusinessError'
            - properties:
                Data:
                  $ref: '#/definitions/api.AccountInfo'
              type: object
        "600":
          description: ""
          schema:
            $ref: '#/definitions/api.BusinessError'
      summary: Account's information
      tags:
      - account
  /accounts/{address}/rewards:
    get:
      consumes:
      - application/json
      description: Query storage rewards for specified account
      parameters:
      - description: The submitter address of the uploaded file
        in: path
        name: address
        type: string
      - default: 0
        description: The number of skipped records, usually it's pageSize * (pageNumber
          - 1)
        in: query
        minimum: 0
        name: skip
        type: integer
      - default: 10
        description: The number of records displayed on the page
        in: query
        maximum: 100
        minimum: 1
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/api.BusinessError'
            - properties:
                Data:
                  $ref: '#/definitions/api.RewardList'
              type: object
        "600":
          description: ""
          schema:
            $ref: '#/definitions/api.BusinessError'
      summary: Account's storage reward list
      tags:
      - account
  /accounts/{address}/txs:
    get:
      consumes:
      - application/json
      description: Query storage transactions for specified account, support root
        hash filter
      parameters:
      - description: The submitter address of the uploaded file
        in: path
        name: address
        type: string
      - default: 0
        description: The number of skipped records, usually it's pageSize * (pageNumber
          - 1)
        in: query
        minimum: 0
        name: skip
        type: integer
      - default: 10
        description: The number of records displayed on the page
        in: query
        maximum: 100
        minimum: 1
        name: limit
        type: integer
      - description: The merkle root hash of the uploaded file
        in: query
        name: rootHash
        type: string
      - description: The layer1 tx hash of the submission
        in: query
        name: txHash
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/api.BusinessError'
            - properties:
                Data:
                  $ref: '#/definitions/api.StorageTxList'
              type: object
        "600":
          description: ""
          schema:
            $ref: '#/definitions/api.BusinessError'
      summary: Account's storage transaction list
      tags:
      - account
  /rewards:
    get:
      consumes:
      - application/json
      description: Query storage rewards
      parameters:
      - default: 0
        description: The number of skipped records, usually it's pageSize * (pageNumber
          - 1)
        in: query
        minimum: 0
        name: skip
        type: integer
      - default: 10
        description: The number of records displayed on the page
        in: query
        maximum: 100
        minimum: 1
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/api.BusinessError'
            - properties:
                Data:
                  $ref: '#/definitions/api.RewardList'
              type: object
        "600":
          description: ""
          schema:
            $ref: '#/definitions/api.BusinessError'
      summary: Storage reward list
      tags:
      - reward
  /stats/address:
    get:
      consumes:
      - application/json
      description: Query hex40 address statistics, including incremental, active and
        full data, and support querying at hourly or daily time intervals
      parameters:
      - default: 0
        description: The number of skipped records, usually it's pageSize * (pageNumber
          - 1)
        in: query
        minimum: 0
        name: skip
        type: integer
      - default: 10
        description: The number of records displayed on the page
        in: query
        maximum: 2000
        minimum: 1
        name: limit
        type: integer
      - description: Timestamp in seconds
        in: query
        name: minTimestamp
        type: integer
      - description: Timestamp in seconds
        in: query
        name: maxTimestamp
        type: integer
      - default: day
        description: Statistics interval
        enum:
        - hour
        - day
        in: query
        name: intervalType
        type: string
      - default: desc
        description: Sort by timestamp
        enum:
        - asc
        - desc
        in: query
        name: sort
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/api.BusinessError'
            - properties:
                Data:
                  $ref: '#/definitions/api.AddressStatList'
              type: object
        "600":
          description: ""
          schema:
            $ref: '#/definitions/api.BusinessError'
      summary: Address statistics
      tags:
      - statistic
  /stats/client:
    get:
      consumes:
      - application/json
      description: Query DA client statistics, including incremental, active and full
        data, and support querying at hourly or daily time intervals
      parameters:
      - default: 0
        description: The number of skipped records, usually it's pageSize * (pageNumber
          - 1)
        in: query
        minimum: 0
        name: skip
        type: integer
      - default: 10
        description: The number of records displayed on the page
        in: query
        maximum: 2000
        minimum: 1
        name: limit
        type: integer
      - description: Timestamp in seconds
        in: query
        name: minTimestamp
        type: integer
      - description: Timestamp in seconds
        in: query
        name: maxTimestamp
        type: integer
      - default: day
        description: Statistics interval
        enum:
        - hour
        - day
        in: query
        name: intervalType
        type: string
      - default: desc
        description: Sort by timestamp
        enum:
        - asc
        - desc
        in: query
        name: sort
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/api.BusinessError'
            - properties:
                Data:
                  $ref: '#/definitions/api.DAClientStatList'
              type: object
        "600":
          description: ""
          schema:
            $ref: '#/definitions/api.BusinessError'
      summary: DA client statistics
      tags:
      - DA statistic
  /stats/fee:
    get:
      consumes:
      - application/json
      description: Query fee statistics, including incremental and full data, and
        support querying at hourly or daily time intervals
      parameters:
      - default: 0
        description: The number of skipped records, usually it's pageSize * (pageNumber
          - 1)
        in: query
        minimum: 0
        name: skip
        type: integer
      - default: 10
        description: The number of records displayed on the page
        in: query
        maximum: 2000
        minimum: 1
        name: limit
        type: integer
      - description: Timestamp in seconds
        in: query
        name: minTimestamp
        type: integer
      - description: Timestamp in seconds
        in: query
        name: maxTimestamp
        type: integer
      - default: day
        description: Statistics interval
        enum:
        - hour
        - day
        in: query
        name: intervalType
        type: string
      - default: desc
        description: Sort by timestamp
        enum:
        - asc
        - desc
        in: query
        name: sort
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/api.BusinessError'
            - properties:
                Data:
                  $ref: '#/definitions/api.FeeStatList'
              type: object
        "600":
          description: ""
          schema:
            $ref: '#/definitions/api.BusinessError'
      summary: Storage fee statistics
      tags:
      - statistic
  /stats/layer1-tx:
    get:
      consumes:
      - application/json
      description: Query transaction statistics, including incremental and full data,
        and support querying at hourly or daily time intervals
      parameters:
      - default: 0
        description: The number of skipped records, usually it's pageSize * (pageNumber
          - 1)
        in: query
        minimum: 0
        name: skip
        type: integer
      - default: 10
        description: The number of records displayed on the page
        in: query
        maximum: 2000
        minimum: 1
        name: limit
        type: integer
      - description: Timestamp in seconds
        in: query
        name: minTimestamp
        type: integer
      - description: Timestamp in seconds
        in: query
        name: maxTimestamp
        type: integer
      - default: day
        description: Statistics interval
        enum:
        - hour
        - day
        in: query
        name: intervalType
        type: string
      - default: desc
        description: Sort by timestamp
        enum:
        - asc
        - desc
        in: query
        name: sort
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/api.BusinessError'
            - properties:
                Data:
                  $ref: '#/definitions/api.TxStatList'
              type: object
        "600":
          description: ""
          schema:
            $ref: '#/definitions/api.BusinessError'
      summary: Layer1 transaction statistics
      tags:
      - statistic
  /stats/miner:
    get:
      consumes:
      - application/json
      description: Query miner statistics, including incremental, active and full
        data, and support querying at hourly or daily time intervals
      parameters:
      - default: 0
        description: The number of skipped records, usually it's pageSize * (pageNumber
          - 1)
        in: query
        minimum: 0
        name: skip
        type: integer
      - default: 10
        description: The number of records displayed on the page
        in: query
        maximum: 2000
        minimum: 1
        name: limit
        type: integer
      - description: Timestamp in seconds
        in: query
        name: minTimestamp
        type: integer
      - description: Timestamp in seconds
        in: query
        name: maxTimestamp
        type: integer
      - default: day
        description: Statistics interval
        enum:
        - hour
        - day
        in: query
        name: intervalType
        type: string
      - default: desc
        description: Sort by timestamp
        enum:
        - asc
        - desc
        in: query
        name: sort
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/api.BusinessError'
            - properties:
                Data:
                  $ref: '#/definitions/api.MinerStatList'
              type: object
        "600":
          description: ""
          schema:
            $ref: '#/definitions/api.BusinessError'
      summary: Miner statistics
      tags:
      - statistic
  /stats/signer:
    get:
      consumes:
      - application/json
      description: Query DA signer statistics, including incremental, active and full
        data, and support querying at hourly or daily time intervals
      parameters:
      - default: 0
        description: The number of skipped records, usually it's pageSize * (pageNumber
          - 1)
        in: query
        minimum: 0
        name: skip
        type: integer
      - default: 10
        description: The number of records displayed on the page
        in: query
        maximum: 2000
        minimum: 1
        name: limit
        type: integer
      - description: Timestamp in seconds
        in: query
        name: minTimestamp
        type: integer
      - description: Timestamp in seconds
        in: query
        name: maxTimestamp
        type: integer
      - default: day
        description: Statistics interval
        enum:
        - hour
        - day
        in: query
        name: intervalType
        type: string
      - default: desc
        description: Sort by timestamp
        enum:
        - asc
        - desc
        in: query
        name: sort
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/api.BusinessError'
            - properties:
                Data:
                  $ref: '#/definitions/api.DASignerStatList'
              type: object
        "600":
          description: ""
          schema:
            $ref: '#/definitions/api.BusinessError'
      summary: DA signer statistics
      tags:
      - DA statistic
...
  /stats/summary:
    get:
      description: Query statistics summary includes `storage fee` and `log sync height`
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/api.BusinessError'
            - properties:
                Data:
                  $ref: '#/definitions/api.Summary'
              type: object
        "600":
          description: ""
          schema:
            $ref: '#/definitions/api.BusinessError'
      summary: Statistics summary
      tags:
      - statistic
...
  /txs/{txSeq}:
    get:
      consumes:
      - application/json
      description: Query storage transaction by txSeq
      parameters:
      - description: storage transaction sequence number
        in: path
        name: txSeq
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/api.BusinessError'
            - properties:
                Data:
                  $ref: '#/definitions/api.StorageTxDetail'
              type: object
        "600":
          description: ""
          schema:
            $ref: '#/definitions/api.BusinessError'
      summary: Storage transaction information
      tags:
      - transaction
swagger: "2.0"